Display shows directly from database
Why: The app contains a page to list all shows, past and present.
Initially, the view uses dummy data, but it needs to be integrated
with the backend.

How: Add query to the appropriate controller in app.py

On branch fyyur
Changes to be committed:
modified:   app.py
modified:   templates/pages/shows.html